Blocks: Extract phrasing content to its own file to allow reuse with React Native
